+++ plugin/src/main/kotlin/ru/cian/huawei/publish/models/request/UpdateReleaseNotesRequest.kt
@@ -1,0 +1,18 @@
+package ru.cian.huawei.publish.models.request
+
+import com.google.gson.annotations.SerializedName
+
+internal data class UpdateReleaseNotesRequest(
+    /**
+     * String(64)
+     * Language. For details, please refer to Languages.
+     */
+    @SerializedName("lang")
+    var lang: String,
+    /**
+     * String(500)
+     * New features in a language.
+     */
+    @SerializedName("newFeatures")
+    var newFeatures: String
+)