+++ plugin/src/main/kotlin/ru/cian/huawei/publish/utils/BuildFileProvider.kt
@@ -5,7 +5,10 @@ import com.android.build.api.variant.ApplicationVariant
 import ru.cian.huawei.publish.BuildFormat
 import java.io.File
 
-internal class BuildFileProvider(private val variant: ApplicationVariant) {
+internal class BuildFileProvider constructor(
+    private val variant: ApplicationVariant,
+    private val logger: Logger,
+) {
 
     fun getBuildFile(buildFormat: BuildFormat): File? {
         return when (buildFormat) {
@@ -18,7 +21,7 @@ internal class BuildFileProvider(private val variant: ApplicationVariant) {
     // TODO(a.mirko): Remove after https://github.com/gradle/gradle/issues/16775
     private fun getFinalApkArtifactCompat(variant: ApplicationVariant): List<File> {
         val apkDirectory = variant.artifacts.get(SingleArtifact.APK).get()
-        Logger.v("Build File Directory: $apkDirectory")
+        logger.v("Build File Directory: $apkDirectory")
         return variant.artifacts.getBuiltArtifactsLoader().load(apkDirectory)
             ?.elements?.map { element -> File(element.outputFile) }
             ?: apkDirectory.asFileTree.matching { include("*.apk") }.map { it.absolutePath }.map { File(it) }