+++ plugin/src/main/kotlin/ru/cian/huawei/publish/utils/FileWrapper.kt
@@ -1,0 +1,9 @@
+package ru.cian.huawei.publish.utils
+
+import java.io.File
+
+class FileWrapper {
+    fun getFile(path: String): File {
+        return File(path)
+    }
+}